Moved files into coredata package
